backend_py/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.8.6
supabase==2.4.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.12.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
httpx==0.25.2
aiofiles==23.2.1
